Added new YAML UI Definitions
- deleted options.js
