+++ jamovi/ancova.u.yaml
@@ -1,0 +1,287 @@
+name:  Ancova
+title: ANCOVA
+jus:   '1.0'
+
+stage: 1
+
+update: ./ancova.events::update
+
+children:
+  - type: variablesupplier
+    name: variablesupplier
+    suggested:
+      - continuous
+      - nominal
+      - ordinal
+    persistentItems: false
+    stretchFactor: 1
+    children:
+      - name: dependent
+        type: variabletargetlistbox
+        label: Dependent Variable
+        showColumnHeaders: false
+        maxItemCount: 1
+        itemDropBehaviour: overwrite
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+      - type: variabletargetlistbox
+        name: fixedFactors
+        label: Fixed Factors
+        showColumnHeaders: false
+        events:
+            change: ./ancova.events::onChange_fixedFactors
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+      - type: variabletargetlistbox
+        name: covariates
+        label: Covariates
+        showColumnHeaders: false
+        events:
+            change: ./ancova.events::onChange_covariates
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+  - type:  label
+    label: Effect Size
+    margin: large
+    style: list-inline
+    children:
+      - type: checkbox
+        name: etaSq
+        label: η²
+      - type: checkbox
+        name: etaSqP
+        label: partial η²
+      - type: checkbox
+        name: omegaSq
+        label: ω²
+
+  - type: collapsebox
+    label: Model
+    collapsed: true
+    stretchFactor: 1
+    children:
+      - type: supplier
+        name: modelSupplier
+        label: Componets
+        persistentItems: true
+        stretchFactor: 1
+        children:
+          - type: targetlistbox
+            name: modelTerms
+            label: Model Terms
+            valueFilter: unique
+            showColumnHeaders: false
+            itemDropBehaviour: emptyspace
+            events:
+                change: ./ancova.events::onChange_modelTerms
+                preprocess: ./ancova.events::onEvent_modelTerms_preprocess
+            columns:
+              - type: listitem.termlabel
+                name: column1
+                label:
+                stretchFactor: 1
+
+      - type: layoutbox
+        children:
+          - type: combobox
+            name: ss
+            label: Sum of squares
+            options:
+              - label: Type 1
+                value: Type 1
+              - label: Type 2
+                value: Type 2
+              - label: Type 3
+                value: Type 3
+
+  - type: collapsebox
+    label: Assumption Checks
+    collapsed: true
+    stretchFactor: 1
+    children:
+      - type: checkbox
+        name: homo
+        label: Homogeneity tests
+
+      - type: checkbox
+        name: qq
+        label: Q-Q plot of residuals
+
+  - type: collapsebox
+    label: Contrasts
+    collapsed: true
+    stretchFactor: 1
+    children:
+      - type: listbox
+        name: contrasts
+        label: Factors
+        showColumnHeaders: false
+        columns:
+          - type: listitem.variablelabel
+            name: var
+            label:
+            selectable: false
+            stretchFactor: 0.5
+          - type: listitem.combobox
+            name: type
+            label:
+            selectable: false
+            stretchFactor: 0.5
+            options:
+                - none
+                - deviation
+                - simple
+                - difference
+                - helmert
+                - repeated
+                - polynomial
+
+  - type: collapsebox
+    label: Post Hoc Tests
+    collapsed: true
+    stretchFactor: 1
+    children:
+      - type: supplier
+        name: postHocSupplier
+        persistentItems: false
+        stretchFactor: 1
+        children:
+          - type: targetlistbox
+            name: postHoc
+            showColumnHeaders: false
+            columns:
+              - type: listitem.termlabel
+                name: column1
+                label:
+                stretchFactor: 1
+
+      - type: label
+        label: Correction
+        children:
+          - type: checkbox
+            name: corrTukey
+            label: Tukey
+
+          - type: checkbox
+            name: corrScheffe
+            label: Scheffe
+
+          - type: checkbox
+            name: corrBonf
+            label: Bonferroni
+
+          - type: checkbox
+            name: corrHolm
+            label: Holm
+
+  - type: collapsebox
+    label: Descriptive Plots
+    collapsed: true
+    stretchFactor: 1
+    children:
+      - type: supplier
+        name: plotsSupplier
+        stretchFactor: 1
+        persistentItems: false
+        children:
+          - type: targetlistbox
+            name: descPlotsHAxis
+            label: Horizontal axis
+            showColumnHeaders: false
+            maxItemCount: 1
+            itemDropBehaviour: overwite
+            columns:
+              - type: listitem.variablelabel
+                name: column1
+                label:
+                stretchFactor: 1
+
+          - type: targetlistbox
+            name: descPlotsSepLines
+            label: Separate lines
+            showColumnHeaders: false
+            maxItemCount: 1
+            itemDropBehaviour: overwite
+            columns:
+              - type: listitem.variablelabel
+                name: column1
+                label:
+                stretchFactor: 1
+
+          - type: targetlistbox
+            name: descPlotsSepPlots
+            label: Separate plots
+            showColumnHeaders: false
+            maxItemCount: 1
+            itemDropBehaviour: overwite
+            columns:
+              - type: listitem.variablelabel
+                name: column1
+                label:
+                stretchFactor: 1
+
+      - type: label
+        label: Display
+        children:
+          - type: radiobutton
+            name: errBarDef_ci
+            optionId: plotError
+            checkedValue: ci
+            label: Confidence interval
+            events:
+                change: ./ancova.events::onChange_errBarDef_ci
+            children:
+              - type: textbox
+                name: ciWidth
+                label: Interval
+                suffix: "%"
+                format: number
+                inputPattern: "[0-9]+"
+
+          - type: radiobutton
+            name: errBarDef_se
+            optionId: plotError
+            checkedValue: se
+            label: Standard Error
+
+  - type: collapsebox
+    label: Additional Options
+    collapsed: true
+    stretchFactor: 1
+    children:
+      - type: checkbox
+        name: compMain
+        label: Compare main effects
+        children:
+          - type: combobox
+            name: compMainCorr
+            label: Correction
+            options:
+              - label: None
+                value: none
+              - label: Tukey
+                value: tukey
+              - label: Bonferroni
+                value: bonferroni
+              - label: Scheffe
+                value: scheffe
+              - label: Sidak
+                value: sidak
+      - type: label
+        label: Display
+        children:
+          - type: checkbox
+            name: descStats
+            label: Descriptive statistics
