+++ jamovi/conttablespaired.u.yaml
@@ -1,0 +1,57 @@
+name:  ContTablesPaired
+title: Paired Samples Contingency Tables
+jus:   '1.0'
+
+children:
+  - type: variablesupplier
+    persistentItems: false
+    stretchFactor: 1
+    children:
+      - type: variabletargetlistbox
+        name: rows
+        label: Rows
+        showColumnHeaders: false
+        maxItemCount: 1
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+      - type: variabletargetlistbox
+        name: cols
+        label: Columns
+        showColumnHeaders: false
+        maxItemCount: 1
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+      - type: variabletargetlistbox
+        name: counts
+        label: Counts (optional)
+        showColumnHeaders: false
+        maxItemCount: 1
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+  - type: layoutbox
+    label: Cells
+    stretchFactor: 1
+    margin: large
+    children:
+      - type: label
+        label: Percentages
+        children:
+          - type: checkbox
+            name: pcRow
+            label: Row
+
+          - type: checkbox
+            name: pcCol
+            label: Column
