+++ jamovi/proptest2.u.yaml
@@ -1,0 +1,76 @@
+name:  PropTest2
+title: Proportion Test (2 Outcomes)
+jus:   '1.0'
+
+update: ui.ciWidth.setEnabled(ui.ci.value());
+
+children:
+  - type: variablesupplier
+    suggested:
+      - continuous
+      - nominal
+      - ordinal
+    persistentItems: false
+    stretchFactor: 1
+    children:
+      - type: variabletargetlistbox
+        name: vars
+        showColumnHeaders: false
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+  - type: layoutbox
+    stretchFactor: 1
+    margin: large
+    children:
+      - type: checkbox
+        name: areCounts
+        label: Values are counts
+
+      - type: layoutbox
+        children:
+          - type: textbox
+            name: testValue
+            label: Test value
+            format: number
+            inputPattern: "[0-9]+"
+
+  - type: label
+    label: Hypothesis
+    children:
+      - type: radiobutton
+        name: hypothesis_notequal
+        label: "≠ Test value"
+        optionId: hypothesis
+        checkedValue: notequal
+
+      - type: radiobutton
+        name: hypothesis_greater
+        label: "> Test value"
+        optionId: hypothesis
+        checkedValue: greater
+
+      - type: radiobutton
+        name: hypothesis_less
+        label: "< Test value"
+        optionId: hypothesis
+        checkedValue: less
+
+  - type: label
+    label: Additional Statistics
+    children:
+      - type: checkbox
+        name: ci
+        label: Confidence intervals
+        events:
+            change: ui.ciWidth.setEnabled(ui.ci.value());
+        children:
+          - type: textbox
+            name: ciWidth
+            label: Interval
+            suffix: "%"
+            format: number
+            inputPattern: "[0-9]+"
