+++ jamovi/ttestones.u.yaml
@@ -1,0 +1,144 @@
+name:  TTestOneS
+title: One Sample T-Test
+jus:   '1.0'
+
+update: ./ttestones.events::update
+
+children:
+  - type: variablesupplier
+    suggested:
+      - continuous
+    permitted:
+      - continuous
+      - nominal
+      - ordinal
+    persistentItems: false
+    stretchFactor: 1
+    children:
+      - type: variabletargetlistbox
+        name: vars
+        label: Dependent Variables
+        showColumnHeaders: false
+        columns:
+          - type: listitem.variablelabel
+            name: column1
+            label:
+            stretchFactor: 1
+
+  - type: layoutbox
+    margin: large
+    stretchFactor: 1
+    children:
+      - type: layoutbox
+        cell: [0, 0]
+        stretchFactor: 1
+        children:
+          - type: label
+            label: Tests
+            children:
+              - type: checkbox
+                name: students
+                label: Student's
+                children:
+                  - type: checkbox
+                    name: bf
+                    label: Bayes factor
+                    events:
+                        change: ./ttestones.events::onChange_bf
+                    children:
+                      - type: textbox
+                        name: bfPrior
+                        label: Prior
+                        format: number
+                        inputPattern: "[0-9]+"
+
+              - type: checkbox
+                name: mann
+                label: Mann-Whitney U
+
+          - type: label
+            label: Hypothesis
+            children:
+              - type: layoutbox
+                children:
+                  - type: textbox
+                    name: testValue
+                    label: Test value
+                    format: number
+                    inputPattern: "[0-9]+"
+
+              - type: radiobutton
+                name: hypothesis_dt
+                optionId: hypothesis
+                checkedValue: dt
+                label: "≠ Test value"
+
+              - type: radiobutton
+                name: hypothesis_gt
+                optionId: hypothesis
+                checkedValue: gt
+                label: "> Test value"
+
+              - type: radiobutton
+                name: hypothesis_lt
+                optionId: hypothesis
+                checkedValue: lt
+                label: "< Test value"
+
+          - type: label
+            label: Missing values
+            children:
+              - type: radiobutton
+                name: miss_perAnalysis
+                optionId: miss
+                checkedValue: perAnalysis
+                label: Exclude cases analysis by analysis
+
+              - type: radiobutton
+                name: miss_listwise
+                optionId: miss
+                checkedValue: listwise
+                label: Exclude cases listwise
+
+      - type: layoutbox
+        cell: [1, 0]
+        stretchFactor: 1
+        children:
+          - type: label
+            label: Additional Statistics
+            children:
+              - type: checkbox
+                name: meanDiff
+                label: Mean difference
+
+              - type: checkbox
+                name: effectSize
+                label: Effect size
+
+              - type: checkbox
+                name: ci
+                label: Confidence interval
+                events:
+                    change: ./ttestones.events::onChange_ci
+                children:
+                  - type: textbox
+                    name: ciWidth
+                    label: Interval
+                    suffix: "%"
+                    format: number
+                    inputPattern: "[0-9]+"
+
+              - type: checkbox
+                name: desc
+                label: Descriptives
+
+              - type: checkbox
+                name: plots
+                label: Descriptives plots
+
+          - type: label
+            label: Assumption Checks
+            children:
+              - type: checkbox
+                name: norm
+                label: Normality
